Fix portal room door cannot open
Same issue with mazes. New design fixes that.
However, the initial design was ambiguous on its intended behaviors:
1. The waterwheels can be activated with the same redstone link frequency to the maze ones. (dispenser + linear chassis)
2. The button in the room activates the sequenced gearshifts, and they should open the door, but then the button also powers a power latch, which powers a redstone link (dispenser + linear chassis) to the waterwheel's dispensers, from my understanding, this deactivates the waterwheels (the dispensers fire again and retract the water) and make them cannot be activated again because the power latch constantly powers the redstone link (dispenser + linear chassis), thus the door cannot be closed after being opened. 
This increases the difficulty of this room because there are deadly spawners in the portal room, and if the door cannot be closed, mobs can get through easily.
3. However, I observed that the sequenced gearshifts also programmed a closing sequence when activate again (although it does not work because water was removed after button press). This conflicts with 2 above.
I cannot decide which one was intended, so I followed 2 above, by fixing the waterwheel power issue, and simply remove the closing sequence on the sequenced gearshift to prevent it closing. If this was not intended, it can be changed later.